requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.4
django-ckeditor==6.2.0
django-cors-headers==3.12.0
django-heroku==0.3.1
django-js-asset==1.2.2
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.8.6
python-decouple==3.6
python-dotenv==0.21.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
whitenoise==5.3.0
